// src/types/transcript.ts
export type CultureKey =
  | 'hi-IN'   // Hindi – India
  | 'bn-IN'   // Bengali – India
  | 'mr-IN'   // Marathi – India
  | 'ta-IN'   // Tamil – India
  | 'te-IN'   // Telugu – India
  | 'en-US'   // English – USA
  | 'en-GB'   // English – UK
  | 'en-AU'   // English – Australia
  | 'es-MX'   // Spanish – Mexico
  | 'es-ES'   // Spanish – Spain
  | 'pt-BR'   // Portuguese – Brazil
  | 'fr-FR'   // French – France
  | 'de-DE'   // German – Germany
  | 'ja-JP'   // Japanese – Japan
  | 'ko-KR'   // Korean – South Korea
  | 'zh-CN'   // Chinese – Mainland
  | 'ar-EG'   // Arabic – Egypt
  | 'ru-RU'   // Russian
  | 'it-IT'   // Italian
  | 'tr-TR'   // Turkish